backend/requirements.txt
django
django-csp
Pillow
djangorestframework
markdown
django-filter
django-cors-headers
psycopg2-binary
whitenoise
pylibmc